--- background.js
@@ -1,8 +1,0 @@
-/**
- * Background Script for Kana Layout Extension
- */
-
-chrome.runtime.onInstalled.addListener(() => {
-  chrome.storage.sync.set({ enabled: true });
-  console.log("Kana Layout Extension installed");
-});
