--- manifest.json
@@ -1,21 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Kana Layout Extension",
-  "version": "1.0.0",
-  "description": "タイピングゲーム用のかな入力配列エミュレーター",
-  "permissions": ["activeTab", "storage"],
-  "content_scripts": [
-    {
-      "matches": ["<all_urls>"],
-      "js": ["content.js"],
-      "run_at": "document_start"
-    }
-  ],
-  "background": {
-    "service_worker": "background.js"
-  },
-  "action": {
-    "default_popup": "popup.html",
-    "default_title": "Kana Layout Extension"
-  }
-}
