+++ src/popup/index.html
@@ -4,7 +4,13 @@
     <meta charset="UTF-8" />
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app">
+      <div class="header">
+        <div class="title">かな配列エミュレーター</div>
+        <div class="subtitle">月配列用</div>
+        <button id="counter" type="button"></button>
+      </div>
+    </div>
     <script type="module" src="./main.js"></script>
   </body>
 </html>
